pyproject.toml
[tool.poetry]
name = "birthday-service"
version = "0.1.0"
description = ""
authors = ["Zhurbenko Maksim <zhurbenkoma97@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-email = "^3.0.0"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
